requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg[binary]==3.2.12
gunicorn==20.1.0
python-dotenv==1.0.0
